Add files via upload
These are EagleCad 7.6.0 library files for Arduino NANO.
There is the original NANO with 0.60in pin width.
Then there are various versions of the V3... The one WITHOUT the 0.91 OLED display is the same as the original NANO.
The versions WITH the 0.91 OLED displays have 2 different pin-width spacings...
Functionally, they are the same, however form and fit are not. 
The LICENSED IdeaSpark version has the standard 1/10th inch spacing 0.90in (22.86mm) pin-width.
The unlicensed knock-off versions have an unusual 22mm pin-width. 
Both forms have a lib with and without the 6-pin ICSP pins so you can employ either external or on-board connections.